requirements.txt
# requirements.txt
fastapi
uvicorn
gunicorn
python-multipart
streamlit
requests
google-generativeai
python-dotenv
pdfplumber
openpyxl
python-docx
pandas
pillow
pytesseract
openai
